backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
asyncpg==0.30.0
pyjwt[crypto]==2.9.0
cryptography==43.0.0
pydantic-settings==2.5.0
typesense==0.21.0
resend==2.0.0
python-multipart==0.0.9
httpx==0.27.0
slowapi==0.1.9
